Welcome to Embedded Innovator questions repository! This repository is a collection of interview questions and resources for various embedded verticals. Whether you're a job seeker preparing for an interview or an interviewer looking for questions to assess candidates, this repository aims to provide a wide range of topics to cover.
This repository serves as a comprehensive resource for anyone interested in embedded systems and related domains.

Whether you are a beginner or an experienced professional, this repository aims to help you prepare for interviews, expand your knowledge, and enhance your understanding of embedded systems.

To access the interview questions and resources in this repository, you can either clone the repository to your local machine or explore the content directly on the GitHub web interface.

To clone the repository, use the following command:

https://github.com/imsunilvaghela/TheEmbeDEADInterview.git

The interview questions in this repository are organized into folders based on the various embedded verticals. Here's an overview of the topics covered:

- C
- CPP
- C+CPP
- Data structures
- Embedded systems
- Linux Kernel(Internals,Debugging,Boot,Process Management,Linux FileSystems)

Feel free to explore the folders and dive into the specific areas that interest you or align with your interview preparation needs.